订单退款退库存测试完善

+++ oms/src/main/java/com/upedge/oms/modules/order/request/OrderRefundItemUpdateRequest.java
@@ -40,7 +40,7 @@ public class OrderRefundItemUpdateRequest{
      */
     private String variantImage;
 
-    public OrderRefundItem toOrderRefundItem(Integer id){
+    public OrderRefundItem toOrderRefundItem(Long id){
         OrderRefundItem orderRefundItem=new OrderRefundItem();
         orderRefundItem.setId(id);
         orderRefundItem.setOrderId(orderId);
